Cree el componente Productos y elimine archivos
