--- src/app/admin/admin.component.html
@@ -1,1 +1,0 @@
-<p>admin works!</p>
