--- src/app/admin/admin.component.ts
@@ -1,19 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-admin',
-  templateUrl: './admin.component.html',
-  styleUrls: ['./admin.component.css']
-})
-export class AdminComponent implements OnInit {
-  
-  modalVisible:boolean=false;
-  adminVisible=false;
-  
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
