--- src/app/model/usuarios.ts
@@ -1,4 +1,0 @@
-export interface Usuarios {
-    nombre: string;
-    contrasena: string;
-}
