--- src/app/productos/aceites/aceites.component.css
@@ -1,51 +1,0 @@
-p-card{
-    width: 21%;
-    margin: 2%;
-    height: 5%;
-    border: solid 2px red;
-    display: inline-block;
-}
-
-p{
-    overflow: hidden;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-}
-
-.p-float-label{
-    margin: 35px;
-}
-
-#float-input{
-    width: 40vw;
-}
-
-img{
-    width: 100%;
-    height: 300px;
-}
-
-p-card button{
-    margin: 0 14.83%;
-}
-
-#botonAgregar{
-    margin-top: 2%;
-    margin-left: 2%;
-    display: block;
-}
-
-#botonAgregarForm{
-    margin-top: 2%;
-    margin-left: 36%;
-    display: block;
-}
-
-#previsualizacion{
- width: 150px;
- height: 180px;
- margin-left: 36%;
- padding: 1%;
- border: dashed 2px blue;
- border-radius: 5px;   
-}
