+++ src/app/servicios/productos.service.ts
@@ -1,31 +1,23 @@
 import { Injectable } from '@angular/core';
-import { AngularFirestore, AngularFirestoreCollection } from '@angular/fire/compat/firestore';
+import { AngularFirestoreCollection, AngularFirestore } from '@angular/fire/compat/firestore';
 import { map } from 'rxjs';
 import { Producto } from '../model/producto';
-
-
-/*import { Injectable } from '@angular/core';
-import { AngularFirestore, AngularFirestoreCollection } from '@angular/fire/compat/firestore';
-import { async, map } from '@firebase/util';
-import { Action } from 'rxjs/internal/scheduler/Action'; */
 
 @Injectable({
   providedIn: 'root'
 })
-export class AceitesService {
-  private coleccionAceites: AngularFirestoreCollection<Producto>;
+export class ProductosService {
+  private coleccionProductos: AngularFirestoreCollection<Producto>;
 
   constructor(private db: AngularFirestore) {
-    this.coleccionAceites = db.collection('aceites');
+    this.coleccionProductos= db.collection('productos');
   }
 
-
-  //Obtenemos todos los datos guardados en la base de dato
+  //Obtenemos todos los datos guardados en la base de datos
   getProductos(){
-    return this.coleccionAceites.snapshotChanges().
+    return this.coleccionProductos.snapshotChanges().
     pipe(map(action=>action.map(a=>a.payload.doc.data())))  
   } 
-
 
   //Creo un producto
   createProducto(nuevoProducto: Producto, url: string) {
@@ -34,7 +26,7 @@ export class AceitesService {
         const id = this.db.createId();
         nuevoProducto.idProducto = id;
         nuevoProducto.img = url
-        const respuesta = await this.coleccionAceites.doc(id).set(nuevoProducto);
+        const respuesta = await this.coleccionProductos.doc(id).set(nuevoProducto);
         resolve(respuesta) 
       } 
       catch (error) { 
@@ -45,14 +37,14 @@ export class AceitesService {
 
   //Edito el producto
   editarProducto(idProducto: string, nuevosDatos: Producto) {
-    return this.coleccionAceites.doc(idProducto).update(nuevosDatos)
+    return this.coleccionProductos.doc(idProducto).update(nuevosDatos)
   }
 
   //Elimino el producto
   deleteProducto(idProducto: string) {
     return new Promise((resolve, reject) => {
       try {
-        const respuesta = this.coleccionAceites.doc(idProducto).delete()
+        const respuesta = this.coleccionProductos.doc(idProducto).delete()
         resolve(respuesta)
       }
       catch (error) {
