+++ src/app/shared/menu/menu.component.html
@@ -1,23 +1,5 @@
 
 <p-menubar [model]="items" style="font-size: 20px;">
-    <button id="boton" type="button" (click)="showModalDialog()" style="margin-top: 15px;" pButton pRipple class="p-button-rounded p-button-success"  label="Iniciar">
+    <button id="boton" type="button" routerLink="login" style="margin-top: 15px;" pButton pRipple class="p-button-rounded p-button-success"  label="Iniciar">
     </button>
 </p-menubar>
-
-<!-- Acá se presenta un formulario para ingresar el usuario y contraseña -->
-<div class="wrapper">
-    <p-dialog header="Iniciar sesion" [(visible)]="modalVisible">
-
-        <form [formGroup]="Usuarios">
-            <!-- Label Nombre Usuario -->
-            <span class="p-float-label">
-                <h2>Usuario</h2>
-                <input type="text" id="float-input" pInputText formControlName="nombre">
-            </span>
-            <!-- Label Contraseña -->
-            <span class="p-float-label">
-                <h2>Contraseña</h2>
-                <input type="password" id="float-input" pInputText formControlName="contrasena">
-            </span>
-            <button id="boton" style="margin-top: 15px; " pButton pRipple type="submit" class="p-button-rounded  p-button-success" (click)="verificarUsuario()" label="Ingresar"></button>
-        </form>
